Bug 1367576 - Make IsWidgetStateSafe not truncate the bits that it needs to test.  r=karlt
This fix a mistake that goes back to the original code from bug 174585
(gecko-dev 9611b23530704402a714fa39cb433a01dca0bb6e, 2005-08-20).

(This makes me wonder how important the code is in the first place if it
didn't work correctly.)

MozReview-Commit-ID: B6q0o5n5hDw

+++ widget/gtk/nsNativeThemeGTK.cpp
@@ -139,7 +139,7 @@ static bool IsWidgetStateSafe(uint8_t* aSafeVector,
                                 GtkWidgetState *aWidgetState)
 {
   MOZ_ASSERT(aWidgetType < ThemeWidgetType_COUNT);
-  uint8_t key = GetWidgetStateKey(aWidgetType, aWidgetState);
+  uint16_t key = GetWidgetStateKey(aWidgetType, aWidgetState);
   return (aSafeVector[key >> 3] & (1 << (key & 7))) != 0;
 }
 
@@ -148,7 +148,7 @@ static void SetWidgetStateSafe(uint8_t *aSafeVector,
                                GtkWidgetState *aWidgetState)
 {
   MOZ_ASSERT(aWidgetType < ThemeWidgetType_COUNT);
-  uint8_t key = GetWidgetStateKey(aWidgetType, aWidgetState);
+  uint16_t key = GetWidgetStateKey(aWidgetType, aWidgetState);
   aSafeVector[key >> 3] |= (1 << (key & 7));
 }
